Exception al borrar colectivo asignado a un viaje

+++ src/main/java/francoceschan/TPJavaBackend/service/colectivo/ColectivoServiceImpl.java
@@ -1,6 +1,7 @@
 package francoceschan.TPJavaBackend.service.colectivo;
 
 import francoceschan.TPJavaBackend.dao.ColectivoDao;
+import francoceschan.TPJavaBackend.dao.ViajeDao;
 import francoceschan.TPJavaBackend.model.Colectivo;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -13,6 +14,7 @@ import java.util.List;
 public class ColectivoServiceImpl implements ColectivoService{
 
     private final ColectivoDao colectivoDao;
+    private final ViajeDao viajeDao;
 
     @Override
     public Colectivo guardarColectivo(Colectivo colectivo){
@@ -26,7 +28,13 @@ public class ColectivoServiceImpl implements ColectivoService{
 
     @Override
     @Transactional
-    public void deleteById(String patente){
+    public void deleteById(String patente) throws RuntimeException {
+
+        // Verifico si hay viajes asociados
+        if (viajeDao.countViajeByColectivo_Patente(patente) > 0) {
+            throw new RuntimeException("El colectivo no se puede eliminar porque está asignado a uno o más viajes");
+        }
+
         colectivoDao.deleteByPatente(patente);
     }
 }
